# always expand ivar arrays to max width