If the instance variable table hasn't been "expanded", allocate the
maximum size of the ivar table.  This operates under the assumption that
most objects will eventually expand their ivar array to the maximum
width anyway, so we may as well avoid realloc calls.