@jhawthorn 2020-01-06
This change seems to make a significant improvement to building `ActionDispatch::Request.new`, which Rails needs to do a few times per-request. This object has quite a few ivars - at least 12, but sometimes more set later, this previously caused 4 reallocs  during each initialization.

## Benchmark

```
# frozen_string_literal: true

require "action_dispatch"
require "benchmark/ips"

REQUEST_ENV = Rack::MockRequest.env_for("http://example.org/")
REQUEST_ENV.freeze

Benchmark.ips do |x|
  x.report "ActionDispatch::Request.new" do
    ActionDispatch::Request.new(REQUEST_ENV)
  end
end
```

**Before:**

```
Warming up --------------------------------------
ActionDispatch::Request.new
                       166.051k i/100ms
Calculating -------------------------------------
ActionDispatch::Request.new
                          2.076M (± 0.5%) i/s -     10.461M in   5.039117s
```


**After**


```
Warming up --------------------------------------
ActionDispatch::Request.new
                       188.776k i/100ms
Calculating -------------------------------------
ActionDispatch::Request.new
                          2.466M (± 0.4%) i/s -     12.459M in   5.051974s
```


## `perf record`/`perf report`

**Before**

<img width="765" alt="Screen Shot 2020-01-06 at 10 20 31" src="https://user-images.githubusercontent.com/131752/71841234-cbfdc680-3073-11ea-90cc-98c7169c124b.png">

**After**

<img width="736" alt="Screen Shot 2020-01-06 at 10 23 04" src="https://user-images.githubusercontent.com/131752/71841231-cacc9980-3073-11ea-8669-345b16ed25f7.png">