@jhawthorn 2020-01-06
A synthetic benchmark which doesn't depend on Rails:

```
def build_ivar_class(n)
  src = n.times.map { |i| "@ivar#{i} = nil" }.join(";")
  eval("class Ivar#{n}; def initialize; #{src}; end; end")
end

TESTS = [1,4,5,7,10,13,17]
TESTS.each {|n| build_ivar_class(n) }

require "benchmark/ips"
Benchmark.ips do |x|
  TESTS.each do |n|
    x.report "with #{n} ivars", "Ivar#{n}.new"
  end
end
```

**Before:**

```
Calculating -------------------------------------
        with 1 ivars      8.741M (± 0.9%) i/s -     43.997M in   5.033814s
        with 4 ivars      6.467M (± 0.6%) i/s -     32.553M in   5.033900s
        with 5 ivars      6.191M (± 0.3%) i/s -     31.320M in   5.059259s
        with 7 ivars      5.203M (± 0.2%) i/s -     26.319M in   5.058245s
       with 10 ivars      4.408M (± 0.7%) i/s -     22.082M in   5.009903s
       with 13 ivars      3.720M (± 1.0%) i/s -     18.762M in   5.044013s
       with 17 ivars      3.258M (± 0.8%) i/s -     16.408M in   5.036510s
```


**After:**

```
Calculating -------------------------------------
        with 1 ivars      8.629M (± 0.2%) i/s -     43.521M in   5.043408s
        with 4 ivars      6.377M (± 0.7%) i/s -     32.098M in   5.033957s
        with 5 ivars      6.225M (± 0.3%) i/s -     31.303M in   5.029014s
        with 7 ivars      5.938M (± 0.3%) i/s -     29.790M in   5.017031s
       with 10 ivars      5.591M (± 0.3%) i/s -     28.116M in   5.028891s
       with 13 ivars      5.288M (± 0.3%) i/s -     26.575M in   5.025391s
       with 17 ivars      4.518M (± 0.2%) i/s -     22.627M in   5.008746s
```

As expected this is better on classes with 7 or more ivars. 7 is the first value which would result in a realloc that this avoids (1-3 ivars are stored inline 4-6 ivars fit in the first malloc).